split generate_doc into different trait
