+++ raphtory/src/vectors/graph_entity.rs
@@ -3,17 +3,12 @@ use crate::{
         api::properties::internal::{TemporalPropertiesOps, TemporalPropertyViewOps},
         graph::{edge::EdgeView, vertex::VertexView},
     },
-    prelude::{EdgeViewOps, GraphViewOps, VertexViewOps},
-    vectors::{entity_id::EntityId, EntityDocument},
+    prelude::{GraphViewOps, VertexViewOps},
 };
 use itertools::{chain, Itertools};
 use std::fmt::Display;
 
 pub trait GraphEntity: Sized {
-    fn generate_doc<T>(&self, template: &T) -> EntityDocument
-    where
-        T: Fn(&Self) -> String;
-
     fn generate_property_list<F, D>(
         &self,
         time_fmt: &F,
@@ -89,25 +84,6 @@ impl<G: GraphViewOps> GraphEntity for VertexView<G> {
         let lines = chain!([min_time, max_time], props);
         lines.intersperse("\n".to_owned()).collect()
     }
-
-    fn generate_doc<T>(&self, template: &T) -> EntityDocument
-    where
-        T: Fn(&Self) -> String,
-    {
-        let raw_content = template(self);
-        let content = match raw_content.char_indices().nth(1000) {
-            Some((index, _)) => (&raw_content[..index]).to_owned(),
-            None => raw_content,
-        };
-        // TODO: allow multi document entities !!!!!
-        // shortened to 1000 (around 250 tokens) to avoid exceeding the max number of tokens,
-        // when embedding but also when inserting documents into prompts
-
-        EntityDocument {
-            id: EntityId::Node { id: self.id() },
-            content,
-        }
-    }
 }
 
 impl<G: GraphViewOps> GraphEntity for EdgeView<G> {
@@ -123,18 +99,5 @@ impl<G: GraphViewOps> GraphEntity for EdgeView<G> {
     {
         // TODO: not needed yet
         "".to_owned()
-    }
-    fn generate_doc<T>(&self, template: &T) -> EntityDocument
-    where
-        T: Fn(&Self) -> String,
-    {
-        let content = template(self);
-        EntityDocument {
-            id: EntityId::Edge {
-                src: self.src().id(),
-                dst: self.dst().id(),
-            },
-            content,
-        }
     }
 }
